Improve ingest algorithm
- Improve automated error recovery
- Use sqlalchemy dialect support for all queries
- Implement ingest validator:
  * Lauch count(*) queries on all tables of ingested database
- Handle JSON error oustide http method
- Add dataserver pod for integration tests
  * Add test dataset with simulated errors
- Improve queue partitioning algorithm
- Simplify DATA_URL env variable management
- Improve example script in charge of database deletion
- Add in2p3 CA certificate inside container image
- Prevent mysql connection time-out, which may occur on long ingest

+++ example/delete_database.sh
@@ -1,16 +1,54 @@
 #!/bin/bash
 # Get the number of remaining elements in ingest chunk queue
 
+# Database names:
+# - cosmoDC2_v1_1_4_image_overlap
+# - dc2_run2_1i_dr1b
+# - dc2_errors
+# - dc2_object_run22i_dr6_wfd_v2_00
+
+# Default password
+# PASSWORD=CHANGEME
+
 set -euxo pipefail
 
 DIR=$(cd "$(dirname "$0")"; pwd -P)
 . "$DIR"/../env.sh
 
-DATABASE="cosmoDC2_v1_1_4_image_overlap"
-PASSWORD=CHANGEME
+usage() {
+  cat << EOD
 
-kubectl exec -it $INSTANCE-ingest-db-0 -- mysql -h localhost -u root -p"$PASSWORD" -e "DELETE FROM qservIngest.task WHERE \`database\` LIKE '$DATABASE';"
+Usage: `basename $0` [options] database password
+
+  Available options:
+    -h          this message
+
+Delete a database from Qserv
+EOD
+}
+
+kind=false
+
+# get the options
+while getopts h c ; do
+    case $c in
+	    h) usage ; exit 0 ;;
+	    \?) usage ; exit 2 ;;
+    esac
+done
+shift `expr $OPTIND - 1`
+
+if [ $# -ne 2 ] ; then
+    usage
+    exit 2
+fi
+
+DATABASE="$1"
+PASSWORD="$2"
+
+
+kubectl exec -it $INSTANCE-ingest-db-0 -- mysql -h localhost -u root -p"$PASSWORD" -e "DELETE FROM qservIngest.chunkfile_queue WHERE \`database\` LIKE '$DATABASE';"
 time kubectl exec -it $INSTANCE-repl-ctl-0 -- curl http://localhost:$REPL_CTL_PORT/ingest/database/"$DATABASE"  \
    -X DELETE -H "Content-Type: application/json" \
-   -d '{"auth_key":""}'
+   -d '{"admin_auth_key":""}'
